+++ Build-Package.ps1
@@ -1,4 +1,4 @@
-param([string]$Version = '1.0.15')
+param([string]$Version = '1.0.16')
 
 $ErrorActionPreference = 'Stop'
 $releaseRoot = [IO.Path]::GetFullPath((Join-Path $PSScriptRoot 'release'))